use outlet on users and albums
