+++ src/app/providers/router/Router.tsx
@@ -8,24 +8,33 @@ import { UserPage } from "../../../pages/users/UserPage";
 import { UserAlbumsPage } from "../../../pages/userAlbums/UserAlbumsPage";
 import { AlbumsPage } from "../../../pages/albums/AlbumsPage";
 import { UserTodosPage } from "../../../pages/userTodos/UserTodosPage";
+import { AlbumPhotosPage } from "../../../pages/photos/AlbumPhotosPage";
 
 export const router = createBrowserRouter([
   {
     path: "/",
     element: <MainLayout />,
-
     children: [
       {
         path: "posts",
         element: <PostsPage />,
       },
       { path: "posts/:id", element: <PostPage /> },
-      { path: "users", element: <UsersPage /> },
-      { path: "users/:id", element: <UserPage /> },
-      { path: "users/:id/posts", element: <UserPostsPage /> },
-      { path: "users/:id/albums", element: <UserAlbumsPage /> },
-      { path: "albums/:id/photos", element: <AlbumsPage /> },
-      { path: "users/:id/todos", element: <UserTodosPage /> },
+      {
+        path: "users",
+        element: <UsersPage />,
+        children: [
+          { path: ":id", element: <UserPage /> },
+          { path: ":id/posts", element: <UserPostsPage /> },
+          { path: ":id/albums", element: <UserAlbumsPage /> },
+          { path: ":id/todos", element: <UserTodosPage /> },
+        ],
+      },
+      {
+        path: "albums",
+        element: <AlbumsPage />,
+        children: [{ path: ":id/photos", element: <AlbumPhotosPage /> }],
+      },
     ],
   },
 ]);
