+++ src/pages/albums/AlbumsPage.tsx
@@ -1,28 +1,23 @@
-import { useParams } from "react-router-dom";
+import { Link, Outlet } from "react-router-dom";
 import { albums } from "../../entities/mocks/albumsMock";
-import { photos } from "../../entities/mocks/photosMock";
 import styles from "./AlbumsPage.module.css";
 
 export const AlbumsPage = () => {
-  const { id } = useParams<{ id: string }>();
-  const albumId = Number(id);
-
-  const album = albums.find((a) => a.id === albumId);
-  const albumPhotos = photos.filter((p) => p.albumId === albumId);
-
-  if (!album) return <h1 className={styles.name}>Album not found</h1>;
-
   return (
     <div className={styles.container}>
-      <h1 className={styles.name}>{album.name}</h1>
+      <h1 className={styles.name}>Albums</h1>
       <div className={styles.photos}>
-        {albumPhotos.map((photo) => (
-          <div key={photo.id}>
-            <p className={styles.title}>{photo.title}</p>
-            <img className={styles.photo} src={photo.url} alt={photo.title} />
-          </div>
+        {albums.map((album) => (
+          <Link
+            key={album.id}
+            className={styles.link}
+            to={`${album.id}/photos`}
+          >
+            {album.name}
+          </Link>
         ))}
       </div>
+      <Outlet />
     </div>
   );
 };
