+++ src/pages/photos/AlbumPhotosPage.module.css
@@ -14,7 +14,14 @@
   gap: 40px;
 }
 
-.link {
-  text-decoration: none;
-  color: var(--link);
+.title {
+  color: var(--text);
+  text-align: center;
+  margin-bottom: 5px;
+}
+
+.photo {
+  border-radius: 20px;
+  max-height: 250px;
+  max-width: 250px;
 }
