+++ src/pages/users/UsersPage.tsx
@@ -1,3 +1,4 @@
+import { Outlet } from "react-router-dom";
 import { UserList } from "../../widgets/usersList/UserList";
 import styles from "./UsersPage.module.css";
 
@@ -8,6 +9,7 @@ export const UsersPage = () => {
       <div className={styles.list}>
         <UserList />
       </div>
+      <Outlet />
     </>
   );
 };
